backend/requirements.txt
Flask
pymongo[srv]==3.12
Flask-PyMongo
python-dotenv
firebase-admin